src/modules/vision/RoadMarkingDect/requirements.txt
actionlib==1.11.13
angles==1.9.11
appdirs==1.4.3
attrs==19.3.0
bondpy==1.8.3
camera-calibration==1.12.23
camera-calibration-parsers==1.11.13
catkin==0.7.18
certifi==2020.4.5.1
chardet==3.0.4
cv-bridge==1.12.8
cycler==0.10.0
Cython==0.29.20
decorator==4.4.2
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
docopt==0.6.2
dynamic-reconfigure==1.5.50
easydict==1.9
fasttext==0.9.2
gazebo-plugins==2.5.19
gazebo-ros==2.5.19
gencpp==0.6.0
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.11
gennodejs==2.0.1
genpy==0.6.7
idna==2.9
image-geometry==1.12.8
imageio==2.8.0
imgaug==0.4.0
importlib-metadata==1.6.0
interactive-markers==1.11.4
kiwisolver==1.1.0
laser-geometry==1.6.4
matplotlib==3.0.3
message-filters==1.12.14
more-itertools==8.2.0
networkx==2.4
numpy==1.18.2
opencv-contrib-python==4.2.0.34
opencv-python==4.2.0.34
opencv-python-headless==4.2.0.34
packaging==20.3
pathlib2==2.3.5
Pillow==7.1.1
pip-review==1.1.0
pipreqs==0.4.10
pkg-resources==0.0.0
pluggy==0.13.1
pluginlib==1.11.3
protobuf==3.11.3
py==1.8.1
pybind11==2.5.0
pyparsing==2.4.7
pytest==5.4.1
python-dateutil==2.8.1
python-qt-binding==0.3.4
pytools==2020.1
PyWavelets==1.1.1
qt-dotgraph==0.3.11
qt-gui==0.3.11
qt-gui-cpp==0.3.11
qt-gui-py-common==0.3.11
requests==2.23.0
resource-retriever==1.12.5
rosbag==1.12.14
rosboost-cfg==1.14.6
rosclean==1.14.6
roscreate==1.14.6
rosgraph==1.12.14
roslaunch==1.12.14
roslib==1.14.6
roslint==0.11.0
roslz4==1.12.14
rosmake==1.14.6
rosmaster==1.12.14
rosmsg==1.12.14
rosnode==1.12.14
rosparam==1.12.14
rospy==1.12.14
rosservice==1.12.14
rostest==1.12.14
rostopic==1.12.14
rosunit==1.14.6
roswtf==1.12.14
rqt-action==0.4.9
rqt-bag==0.4.12
rqt-bag-plugins==0.4.12
rqt-console==0.4.8
rqt-dep==0.4.9
rqt-graph==0.4.9
rqt-gui==0.5.0
rqt-gui-py==0.5.0
rqt-image-view==0.4.13
rqt-launch==0.4.8
rqt-logger-level==0.4.8
rqt-moveit==0.5.7
rqt-msg==0.4.8
rqt-nav-view==0.5.7
rqt-plot==0.4.8
rqt-pose-view==0.5.8
rqt-publisher==0.4.8
rqt-py-common==0.5.0
rqt-py-console==0.4.8
rqt-reconfigure==0.5.0
rqt-robot-dashboard==0.5.7
rqt-robot-monitor==0.5.8
rqt-robot-steering==0.5.9
rqt-runtime-monitor==0.5.7
rqt-rviz==0.5.10
rqt-service-caller==0.4.8
rqt-shell==0.4.9
rqt-srv==0.4.8
rqt-tf-tree==0.6.0
rqt-top==0.4.8
rqt-topic==0.4.10
rqt-web==0.4.8
rviz==1.12.17
scikit-image==0.15.0
scipy==1.4.1
sensor-msgs==1.12.7
Shapely==1.7.0
six==1.14.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
tensorboardX==2.0
tf==1.11.9
tf-conversions==1.11.9
tf2-geometry-msgs==0.5.20
tf2-kdl==0.5.20
tf2-py==0.5.20
tf2-ros==0.5.20
topic-tools==1.12.14
torch==1.4.0
torchvision==0.5.0
urllib3==1.25.8
wcwidth==0.1.9
xacro==1.11.3
yarg==0.1.9
zerocm==0.0.0
zipp==1.2.0
